+++ src/controllers/controller.py
@@ -1,14 +1,14 @@
 from PySide6.QtCore import (QObject, Slot, QSettings)
 from PySide6.QtWidgets import (QApplication, QFileDialog)
 
-from view import MainWindow
-from model import Watcher
+from src.view.main_view import MainWindow
+from src.model import Watcher
 
 from time import sleep
 from threading import Thread
 
-from network import (MetaData)
-from .notification_icon import (NotificationIconController)
+from src.network import (MetaData)
+from .notification_controller import (NotificationController)
 
 
 class Controller(QObject):
@@ -41,7 +41,7 @@ class Controller(QObject):
         self.view.show()
 
         # Non so se ci vada il parent su Notification...
-        self.notification_icon = NotificationIconController(app, parent)
+        self.notification_icon = NotificationController(app, parent)
         self.notification_icon.Sg_show_app.connect(lambda: self.view.show())
 
         # Attivo il watchdog nella root definita dall'utente
