+++ src/view/__init__.py
@@ -1,5 +1,16 @@
-from .mainwindow import MainWindow
+from .main_view import MainWindow
+from .main_widget import MainWidget
+from .file_syncronized_widget import FileSyncronizedWidget
+from .lateral_menu_widget import LateralMenuWidget
+from .settings_widget import SettingsWidget
+from .notification_view import NotificationView
+
 
 __all__ = [
-    "MainWindow"
+    "MainWindow",
+    "MainWidget",
+    "FileSyncronizedWidget",
+    "LateralMenuWidget",
+    "SettingsWidget",
+    "NotificationView"
 ]
