+++ src/view/file_syncronized_widget.py
@@ -2,18 +2,18 @@ from PySide6.QtCore import (QSettings, QUrl)
 from PySide6.QtWidgets import (
     QVBoxLayout, QHBoxLayout, QWidget, QScrollArea, QPushButton)
 from PySide6.QtGui import (QDesktopServices)
-from model import Directory
-from view.widgets.subwidget.filewidget import FileWidget
-from view.layouts.flowlayout import FlowLayout
+from src.model import Directory
+from src.view.widgets import FileWidget
+from src.view.layouts.flowlayout import FlowLayout
 
 
-class SyncronizedWidget(QWidget):
+class FileSyncronizedWidget(QWidget):
 
     # creating Signals
     # TODO
 
     def __init__(self, parent=None):
-        super(SyncronizedWidget, self).__init__(parent)
+        super(FileSyncronizedWidget, self).__init__(parent)
 
         self.env_settings = QSettings()
         self.current_dir = Directory('', self.env_settings.value("sync_path"))
