+++ src/view/widgets/settings/__init__.py
@@ -1,10 +1,8 @@
-from .settingswidget import SettingsWidget
 from .set_quota_disk_view import SetQuotaDiskView
 from .set_policy_view import SetPolicyView
 from .set_path_view import SetPathView
 
 __all__ = [
-    "SettingsWidget",
     "SetQuotaDiskView",
     "SetPolicyView",
     "SetPathView"
